Fix bug in Session.php
Without the try/catch block around the calls to $this->browser->target(), if one of the calls throws an exception, the final line that closes $this->browser won't be called. Then, when the parent script completes and the browser object is destructed, an exception is thrown from the destructor in DevtoolsClient with no way to catch it. Adding the try/catch block allows us to close the browser object even with an earlier failure, preventing the later exception.